Float fixed 110/114

+++ srcs/apply_specific_format.c
@@ -72,7 +72,7 @@ void	apply_format_ptr(t_flist *flist, va_list *valist)
 }
 void	apply_format_float(t_flist *flist, va_list *valist)
 {
-	double arg;
+	long double arg;
 
 	if (ft_strequ(flist->length, "L"))
 		arg = va_arg(*valist, long double);
